3rd Wheel and additional outcomes are now active
1. The 3rd Wheel is now active for outcomes.
2. All outcomes are now active (can only see this through the Unity Console at the moment)
3. Lowered the cap on the number of possible rotations for Wheels 1 and 2.

KNOWN ISSUE
The game is currently not able to find SpinWheel3Sprite and MultiWheelSprite, so Wheel 2 will always stop on the wrong icon when Wheel 3 is about to start spinning, and Wheel 3 will stop on the wrong icon for the outcome that was intended if MultiWheel was the payout outcome (this is mostly only noticeable in the Unity Console debug log). I'm assuming BackDoorSprite is also not being recognized for some reason.